Avoid prototype pollution
GraphQL checks reported a possible prototype pollution with the former
object style store of subscribers. This new solution uses the Map class,
which should be generally more robust. This also lead to the the fix of
possible async code issues due to an await inside the higher order
function forEach when raising events.